fix lợi nhuận trong phiếu nhập kho

+++ PTTKHT_BackEnd/src/main/java/PTTKHT_BackEnd/Entity/PhieuNhapKho/PhieuNhapKho.java
@@ -33,7 +33,7 @@ public class PhieuNhapKho implements Serializable {
     private Integer tongGiaTri;
 
     @Column(name = "LoiNhuan", nullable = false)
-    private Integer loiNhuan;
+    private Float loiNhuan;
 
     @ManyToOne
     @JoinColumn(name = "MaNCC", nullable = false)
